# build/swagger.json
{
	"components": {
...
		"schemas": {
			"ErrorStore": {
				"properties": {
					"ErrorMessage": {
						"type": "string"
					}
				},
				"required": [
					"ErrorMessage"
				],
				"type": "object"
			},
...
			"%24Result.DefaultSelection_Prisma.%24StudentPayload_": {
				"properties": {
					"dueAmount": {
						"type": "number",
						"format": "double"
					},
					"subjectsTaken": {
						"items": {
							"type": "string"
						},
						"type": "array"
					},
					"isDeleted": {
						"type": "boolean"
					},
					"whatsappNo": {
						"type": "string"
					},
					"parentPhnNo": {
						"type": "string"
					},
					"parentName": {
						"type": "string"
					},
					"timing": {
						"type": "string"
					},
					"joiningDate": {
						"type": "string",
						"format": "date-time"
					},
					"pastScore": {
						"type": "number",
						"format": "double"
					},
					"studentStatus": {
						"type": "string"
					},
					"boardType": {
						"type": "string"
					},
					"schoolName": {
						"type": "string"
					},
					"classNo": {
						"type": "string"
					},
					"name": {
						"type": "string"
					},
					"id": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"dueAmount",
					"subjectsTaken",
					"isDeleted",
					"whatsappNo",
					"parentPhnNo",
					"parentName",
					"timing",
					"joiningDate",
					"pastScore",
					"studentStatus",
					"boardType",
					"schoolName",
					"classNo",
					"name",
					"id"
				],
				"type": "object"
			},
			"Student": {
				"$ref": "#/components/schemas/%24Result.DefaultSelection_Prisma.%24StudentPayload_",
				"description": "Model Student"
			},
			"PageResponse_Student_": {
				"properties": {
					"items": {
						"items": {
							"$ref": "#/components/schemas/Student"
						},
						"type": "array"
					},
					"totalCount": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"items",
					"totalCount"
				],
				"type": "object"
			},
			"StudentFeeDataInterface": {
				"properties": {
					"feeDetails": {
						"properties": {
							"monthOfPaid": {
								"type": "number",
								"format": "double"
							},
							"dateOfPaid": {
								"type": "string",
								"format": "date-time"
							},
							"paidAmount": {
								"type": "number",
								"format": "double"
							},
							"subjectsTaken": {
								"items": {
									"type": "string"
								},
								"type": "array"
							},
							"id": {
								"type": "number",
								"format": "double"
							}
						},
						"required": [
							"monthOfPaid",
							"dateOfPaid",
							"paidAmount",
							"subjectsTaken",
							"id"
						],
						"type": "object"
					},
					"feeCharge": {
						"properties": {
							"chargedMonth": {
								"type": "number",
								"format": "double"
							},
							"amount": {
								"type": "number",
								"format": "double"
							},
							"id": {
								"type": "number",
								"format": "double"
							}
						},
						"required": [
							"chargedMonth",
							"amount",
							"id"
						],
						"type": "object"
					},
					"parentName": {
						"type": "string"
					},
					"name": {
						"type": "string"
					},
					"id": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"feeDetails",
					"feeCharge",
					"parentName",
					"name",
					"id"
				],
				"type": "object"
			},
			"Pagination": {
				"properties": {
					"count": {
						"type": "number",
						"format": "double"
					},
					"pageSize": {
						"type": "number",
						"format": "double"
					},
					"curentPage": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"count",
					"pageSize",
					"curentPage"
				],
				"type": "object"
			},
			"Pageresponse_StudentFeeDataInterface_": {
				"properties": {
					"pagination": {
						"$ref": "#/components/schemas/Pagination"
					},
					"items": {
						"items": {
							"$ref": "#/components/schemas/StudentFeeDataInterface"
						},
						"type": "array"
					},
					"totalCount": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"pagination",
					"items",
					"totalCount"
				],
				"type": "object"
			},
			"ServiceResponse": {
				"properties": {
					"data": {},
					"response": {
						"type": "string"
					}
				},
				"required": [
					"data",
					"response"
				],
				"type": "object"
			},
			"studentCreateInterface": {
				"properties": {
					"feeCharge": {},
					"dueAmount": {
						"type": "number",
						"format": "double"
					},
					"isDeleted": {
						"type": "boolean"
					},
					"whatsappNo": {
						"type": "string"
					},
					"parentPhnNo": {
						"type": "string"
					},
					"parentName": {
						"type": "string"
					},
					"timing": {
						"type": "string"
					},
					"joiningDate": {
						"type": "string",
						"format": "date-time"
					},
					"subjectsTaken": {
						"items": {
							"type": "string"
						},
						"type": "array"
					},
					"pastScore": {
						"type": "number",
						"format": "double"
					},
					"studentStatus": {
						"type": "string"
					},
					"boardType": {
						"type": "string"
					},
					"schoolName": {
						"type": "string"
					},
					"classNo": {
						"type": "string"
					},
					"name": {
						"type": "string"
					}
				},
				"required": [
					"feeCharge",
					"dueAmount",
					"isDeleted",
					"whatsappNo",
					"parentPhnNo",
					"parentName",
					"timing",
					"joiningDate",
					"subjectsTaken",
					"pastScore",
					"studentStatus",
					"boardType",
					"schoolName",
					"classNo",
					"name"
				],
				"type": "object"
			},
			"studentFeeInterface": {
				"properties": {
					"info": {
						"properties": {
							"name": {
								"type": "string"
							},
							"modeOfPayment": {
								"type": "string"
							},
							"dueAmount": {
								"type": "number",
								"format": "double"
							},
							"id": {
								"type": "number",
								"format": "double"
							}
						},
						"required": [
							"name",
							"modeOfPayment",
							"dueAmount",
							"id"
						],
						"type": "object"
					},
					"data": {
						"properties": {
							"subjectsTaken": {
								"items": {
									"type": "string"
								},
								"type": "array"
							},
							"studentId": {
								"type": "number",
								"format": "double"
							},
							"paidAmount": {
								"type": "number",
								"format": "double"
							},
							"dateOfPaid": {
								"type": "string",
								"format": "date-time"
							}
						},
						"required": [
							"subjectsTaken",
							"studentId",
							"paidAmount",
							"dateOfPaid"
						],
						"type": "object"
					}
				},
				"required": [
					"info",
					"data"
				],
				"type": "object"
			},
			"studentUpdateInterface": {
				"properties": {
					"dueAmount": {
						"type": "number",
						"format": "double"
					},
					"isDeleted": {
						"type": "boolean"
					},
					"whatsappNo": {
						"type": "string"
					},
					"parentPhnNo": {
						"type": "string"
					},
					"parentName": {
						"type": "string"
					},
					"timing": {
						"type": "string"
					},
					"joiningDate": {
						"type": "string",
						"format": "date-time"
					},
					"subjectsTaken": {
						"items": {
							"type": "string"
						},
						"type": "array"
					},
					"pastScore": {
						"type": "number",
						"format": "double"
					},
					"studentStatus": {
						"type": "string"
					},
					"boardType": {
						"type": "string"
					},
					"schoolName": {
						"type": "string"
					},
					"classNo": {
						"type": "string"
					},
					"name": {
						"type": "string"
					}
				},
				"type": "object"
			},
			"%24Result.DefaultSelection_Prisma.%24TransactionsListPayload_": {
				"properties": {
					"transactionMode": {
						"type": "string"
					},
					"modeOfPayment": {
						"type": "string"
					},
					"amount": {
						"type": "number",
						"format": "double"
					},
					"dateOfPayment": {
						"type": "string",
						"format": "date-time"
					},
					"category": {
						"type": "string"
					},
					"description": {
						"type": "string"
					},
					"itemName": {
						"type": "string"
					},
					"id": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"transactionMode",
					"modeOfPayment",
					"amount",
					"dateOfPayment",
					"category",
					"description",
					"itemName",
					"id"
				],
				"type": "object"
			},
			"TransactionsList": {
				"$ref": "#/components/schemas/%24Result.DefaultSelection_Prisma.%24TransactionsListPayload_",
				"description": "Model TransactionsList"
			},
			"PageResponse_TransactionsList_": {
				"properties": {
					"items": {
						"items": {
							"$ref": "#/components/schemas/TransactionsList"
						},
						"type": "array"
					},
					"totalCount": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"items",
					"totalCount"
				],
				"type": "object"
			},
			"transactionlistInterface": {
				"properties": {
					"transactionMode": {
						"type": "string"
					},
					"modeOfPayment": {
						"type": "string"
					},
					"amount": {
						"type": "number",
						"format": "double"
					},
					"dateOfPayment": {
						"type": "string",
						"format": "date-time"
					},
					"category": {
						"type": "string"
					},
					"description": {
						"type": "string"
					},
					"itemName": {
						"type": "string"
					}
				},
				"required": [
					"transactionMode",
					"modeOfPayment",
					"amount",
					"dateOfPayment",
					"category",
					"description",
					"itemName"
				],
				"type": "object"
			},
			"transactionlistUpdateInterface": {
				"properties": {
					"transactionMode": {
						"type": "string"
					},
					"modeOfPayment": {
						"type": "string"
					},
					"amount": {
						"type": "number",
						"format": "double"
					},
					"dateOfPayment": {
						"type": "string",
						"format": "date-time"
					},
					"category": {
						"type": "string"
					},
					"description": {
						"type": "string"
					},
					"itemName": {
						"type": "string"
					}
				},
				"type": "object"
			},
			"TransactionCategory": {
				"properties": {
					"debit": {
						"type": "number",
						"format": "double"
					},
					"credit": {
						"type": "number",
						"format": "double"
					},
					"category": {
						"type": "string"
					}
				},
				"required": [
					"debit",
					"credit",
					"category"
				],
				"type": "object"
			}
		},
		"securitySchemes": {}
	},
	"info": {
		"title": "aarshabackend",
		"version": "1.0.0",
		"license": {
			"name": "MIT"
		},
		"contact": {
			"name": "KranthiNakka-mls ",
			"email": "nakka.kranthi@metrolabsservices.com"
		}
	},
	"openapi": "3.0.0",
	"paths": {
...
		"/student/all/items": {
			"get": {
				"operationId": "GetAllStudents",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"anyOf": [
										{
											"$ref": "#/components/schemas/PageResponse_Student_"
										},
										{
											"$ref": "#/components/schemas/ErrorStore"
										}
									]
								}
							}
						}
					}
				},
				"tags": [
					"Student Management"
				],
				"security": [],
				"parameters": [
					{
						"in": "query",
						"name": "Pagination",
						"required": false,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "Filters",
						"required": false,
						"schema": {
							"type": "array",
							"items": {}
						}
					}
				]
			}
		},
		"/student/feedetails": {
			"get": {
				"operationId": "GetAllFeeDeatails",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"anyOf": [
										{
											"$ref": "#/components/schemas/Pageresponse_StudentFeeDataInterface_"
										},
										{
											"$ref": "#/components/schemas/ErrorStore"
										}
									]
								}
							}
						}
					}
				},
				"tags": [
					"Student Management"
				],
				"security": [],
				"parameters": [
					{
						"in": "query",
						"name": "PageData",
						"required": false,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "Filters",
						"required": false,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/student/getbyname": {
			"post": {
				"operationId": "GetStudentByName",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"anyOf": [
										{},
										{
											"$ref": "#/components/schemas/ErrorStore"
										}
									]
								}
							}
						}
					}
				},
				"tags": [
					"Student Management"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"properties": {
									"searchIput": {
										"type": "string"
									}
								},
								"required": [
									"searchIput"
								],
								"type": "object"
							}
						}
					}
				}
			}
		},
		"/student/{id}": {
			"get": {
				"operationId": "GetStudentById",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"anyOf": [
										{
											"$ref": "#/components/schemas/Student"
										},
										{
											"$ref": "#/components/schemas/ErrorStore"
										}
									]
								}
							}
						}
					}
				},
				"tags": [
					"Student Management"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			},
			"put": {
				"operationId": "UpdateStudentById",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"anyOf": [
										{
											"type": "string"
										},
										{
											"$ref": "#/components/schemas/ErrorStore"
										}
									]
								}
							}
						}
					}
				},
				"tags": [
					"Student Management"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/studentUpdateInterface"
							}
						}
					}
				}
			},
			"delete": {
				"operationId": "DeleteStudentById",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"anyOf": [
										{
											"type": "string"
										},
										{
											"$ref": "#/components/schemas/ErrorStore"
										}
									]
								}
							}
						}
					}
				},
				"tags": [
					"Student Management"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
		"/student/create": {
			"post": {
				"operationId": "CreateStudent",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"anyOf": [
										{
											"$ref": "#/components/schemas/ServiceResponse"
										},
										{
											"$ref": "#/components/schemas/ErrorStore"
										},
										{}
									]
								}
							}
						}
					}
				},
				"tags": [
					"Student Management"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/studentCreateInterface"
							}
						}
					}
				}
			}
		},
		"/student/newfee": {
			"post": {
				"operationId": "AddStudentFee",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"anyOf": [
										{
											"type": "string"
										},
										{
											"$ref": "#/components/schemas/ErrorStore"
										},
										{}
									]
								}
							}
						}
					}
				},
				"tags": [
					"Student Management"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/studentFeeInterface"
							}
						}
					}
				}
			}
		},
		"/transaction/all/items": {
			"get": {
				"operationId": "GetAllTransaction",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"anyOf": [
										{
											"$ref": "#/components/schemas/PageResponse_TransactionsList_"
										},
										{
											"$ref": "#/components/schemas/ErrorStore"
										}
									]
								}
							}
						}
					}
				},
				"tags": [
					"Expanses Transactions"
				],
				"security": [],
				"parameters": [
					{
						"in": "query",
						"name": "Pagination",
						"required": false,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "Filters",
						"required": false,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/transaction/{id}": {
			"get": {
				"operationId": "GetTransactionById",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"anyOf": [
										{
											"$ref": "#/components/schemas/TransactionsList"
										},
										{
											"$ref": "#/components/schemas/ErrorStore"
										}
									]
								}
							}
						}
					}
				},
				"tags": [
					"Expanses Transactions"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			},
			"put": {
				"operationId": "UpdateTransactionById",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"anyOf": [
										{
											"type": "string"
										},
										{
											"$ref": "#/components/schemas/ErrorStore"
										}
									]
								}
							}
						}
					}
				},
				"tags": [
					"Expanses Transactions"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/transactionlistUpdateInterface"
							}
						}
					}
				}
			},
			"delete": {
				"operationId": "DeleteStudentById",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"anyOf": [
										{
											"type": "string"
										},
										{
											"$ref": "#/components/schemas/ErrorStore"
										}
									]
								}
							}
						}
					}
				},
				"tags": [
					"Expanses Transactions"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
		"/transaction/create": {
			"post": {
				"operationId": "CreateStudent",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"anyOf": [
										{
											"$ref": "#/components/schemas/ServiceResponse"
										},
										{
											"$ref": "#/components/schemas/ErrorStore"
										}
									]
								}
							}
						}
					}
				},
				"tags": [
					"Expanses Transactions"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/transactionlistInterface"
							}
						}
					}
				}
			}
		},
		"/transaction/chartdata": {
			"post": {
				"operationId": "PieChartTransactionData",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"anyOf": [
										{
											"items": {
												"$ref": "#/components/schemas/TransactionCategory"
											},
											"type": "array"
										},
										{
											"$ref": "#/components/schemas/ErrorStore"
										},
										{}
									]
								}
							}
						}
					}
				},
				"tags": [
					"Expanses Transactions"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"properties": {
									"to": {
										"type": "string"
									},
									"from": {
										"type": "string"
									}
								},
								"required": [
									"to",
									"from"
								],
								"type": "object"
							}
						}
					}
				}
			}
		}
	},
	"servers": [
		{
			"url": "/"
		}
	]
}